+++ frontend/js/store.js
@@ -1,4 +1,5 @@
 import { Account } from './helpers/Account';
+import { Renderer } from './renders.js';
 
 /**
  * This class hosts all data.
@@ -11,8 +12,17 @@ export class Store {
    */
   #accounts;
 
-  constructor() {
+  /**
+   * @type {Renderer}
+   */
+  #renderer;
+
+  /**
+   * @param {Renderer} renderer
+   */
+  constructor(renderer) {
     this.#accounts = [];
+    this.#renderer = renderer;
   }
 
   /**
@@ -23,7 +33,7 @@ export class Store {
    */
   setAccounts(accounts) {
     this.#accounts = accounts;
-    accountRenderer(this.#accounts);
+    this.#renderer.accountRenderer(this.#accounts);
   }
 
   /**
@@ -32,53 +42,6 @@ export class Store {
    */
   addAccount(account) {
     this.#accounts.push(account);
-    accountRenderer(this.#accounts);
+    this.#renderer.accountRenderer(this.#accounts);
   }
 }
-
-/**
- * Render the account data.
- * @param {Account[]} accounts
- */
-const accountRenderer = (accounts) => {
-  // generate option tags
-  const optionTags = () =>
-    accounts.map((account) => generateOptionTag(account.id, account.username));
-
-  // grab the account select tag from the DOM.
-  const accountFilterSelectTag = $('#account-filter');
-  accountFilterSelectTag.empty(); // clear the select tag.
-  // generate default option tag.
-  const accountNotSelected = generateOptionTag(0, 'ALL');
-  accountFilterSelectTag.append([accountNotSelected, ...optionTags()]);
-
-  // grab the account select tag from the DOM.
-  const accountSelectTag = $('#account');
-  accountSelectTag.empty(); // clear the select tag.
-  // add option tags to the select tag.
-  accountSelectTag.append(optionTags());
-
-  // grab the 'from' select tag from the DOM.
-  const fromSelectTag = $('#from');
-  fromSelectTag.empty(); // clear the select tag.
-  fromSelectTag.append(optionTags());
-
-  // grab the 'to' select tag from the DOM.
-  const toSelectTag = $('#to');
-  toSelectTag.empty(); // clear the select tag.
-  toSelectTag.append(optionTags);
-};
-
-/**
- * generate option tags for the select tag.
- * @param {number} id
- * @param {string} name
- * @returns {HTMLOptionElement} optionTag
- */
-// TODO: to private after the category is updated.
-export const generateOptionTag = (id, name) => {
-  const optionTag = document.createElement('option');
-  optionTag.value = id.toString();
-  optionTag.innerText = name;
-  return optionTag;
-};
